+++ notifiers/webhook_test.go
@@ -2,6 +2,7 @@ package notifiers
 
 import (
 	"github.com/statping/statping/database"
+	"github.com/statping/statping/types/core"
 	"github.com/statping/statping/types/failures"
 	"github.com/statping/statping/types/notifications"
 	"github.com/statping/statping/types/null"
@@ -27,6 +28,7 @@ func TestWebhookNotifier(t *testing.T) {
 	require.Nil(t, err)
 	db.AutoMigrate(&notifications.Notification{})
 	notifications.SetDB(db)
+	core.Example()
 
 	t.Run("Load webhooker", func(t *testing.T) {
 		Webhook.Host = webhookTestUrl
@@ -44,6 +46,11 @@ func TestWebhookNotifier(t *testing.T) {
 
 	t.Run("webhooker Notifier Tester", func(t *testing.T) {
 		assert.True(t, Webhook.CanSend())
+	})
+
+	t.Run("webhooker OnSave", func(t *testing.T) {
+		_, err := Webhook.OnSave()
+		assert.Nil(t, err)
 	})
 
 	t.Run("webhooker OnFailure", func(t *testing.T) {